+++ tests/DependencyInjection.Lifetime.Analyzers.Tests/CodeFixes/DI002_ScopeEscapeCodeFixTests.cs
@@ -170,6 +170,90 @@ public class DI002_ScopeEscapeCodeFixTests
             .VerifyCodeFixAsync(source, expected, fixedSource, "DI002_Suppress");
     }
 
+    [Fact]
+    public async Task CodeFix_AddsPragmaSuppress_AliasedReturnEscape()
+    {
+        var source = Usings + """
+            public interface IMyService { }
+
+            public class MyClass
+            {
+                private readonly IServiceScopeFactory _scopeFactory;
+
+                public MyClass(IServiceScopeFactory scopeFactory)
+                {
+                    _scopeFactory = scopeFactory;
+                }
+
+                public IMyService GetService()
+                {
+                    IMyService escaped;
+                    using (var scope = _scopeFactory.CreateScope())
+                    {
+                        escaped = {|#0:scope.ServiceProvider.GetRequiredService<IMyService>()|};
+                    }
+
+                    return escaped;
+                }
+            }
+
+            public class Startup
+            {
+                public void ConfigureServices(IServiceCollection services)
+                {
+                    services.AddScoped<IMyService, ScopedMyService>();
+                }
+            }
+
+            public class ScopedMyService : IMyService { }
+            """;
+
+        var fixedSource = Usings + """
+            public interface IMyService { }
+
+            public class MyClass
+            {
+                private readonly IServiceScopeFactory _scopeFactory;
+
+                public MyClass(IServiceScopeFactory scopeFactory)
+                {
+                    _scopeFactory = scopeFactory;
+                }
+
+                public IMyService GetService()
+                {
+                    IMyService escaped;
+                    using (var scope = _scopeFactory.CreateScope())
+                    {
+            #pragma warning disable DI002
+                        escaped = scope.ServiceProvider.GetRequiredService<IMyService>();
+            #pragma warning restore DI002
+                    }
+
+                    return escaped;
+                }
+            }
+
+            public class Startup
+            {
+                public void ConfigureServices(IServiceCollection services)
+                {
+                    services.AddScoped<IMyService, ScopedMyService>();
+                }
+            }
+
+            public class ScopedMyService : IMyService { }
+            """;
+
+        var expected = CodeFixVerifier<DI002_ScopeEscapeAnalyzer, DI002_ScopeEscapeCodeFixProvider>
+            .Diagnostic(DiagnosticDescriptors.ScopedServiceEscapes)
+            .WithLocation(0)
+            .WithArguments("return");
+
+        await CodeFixVerifier<DI002_ScopeEscapeAnalyzer, DI002_ScopeEscapeCodeFixProvider>
+            .VerifyCodeFixAsync(source, expected, fixedSource, "DI002_Suppress");
+    }
+
     #endregion
 
     #region Property Assignment Sink
@@ -245,6 +329,184 @@ public class DI002_ScopeEscapeCodeFixTests
             .Diagnostic(DiagnosticDescriptors.ScopedServiceEscapes)
             .WithLocation(19, 19)
             .WithArguments("Service");
+
+        await CodeFixVerifier<DI002_ScopeEscapeAnalyzer, DI002_ScopeEscapeCodeFixProvider>
+            .VerifyCodeFixAsync(source, expected, fixedSource, "DI002_Suppress");
+    }
+
+    #endregion
+
+    #region Ref/Out and Delegate Sinks
+
+    [Fact]
+    public async Task CodeFix_AddsPragmaSuppress_RefParameterAliasEscape()
+    {
+        var source = Usings + """
+            public interface IMyService { }
+
+            public class MyClass
+            {
+                private readonly IServiceScopeFactory _scopeFactory;
+
+                public MyClass(IServiceScopeFactory scopeFactory)
+                {
+                    _scopeFactory = scopeFactory;
+                }
+
+                public void Resolve(ref IMyService escaped)
+                {
+                    using var scope = _scopeFactory.CreateScope();
+                    var service = {|#0:scope.ServiceProvider.GetRequiredService<IMyService>()|};
+                    escaped = service;
+                }
+            }
+
+            public class Startup
+            {
+                public void ConfigureServices(IServiceCollection services)
+                {
+                    services.AddScoped<IMyService, ScopedMyService>();
+                }
+            }
+
+            public class ScopedMyService : IMyService { }
+            """;
+
+        var fixedSource = Usings + """
+            public interface IMyService { }
+
+            public class MyClass
+            {
+                private readonly IServiceScopeFactory _scopeFactory;
+
+                public MyClass(IServiceScopeFactory scopeFactory)
+                {
+                    _scopeFactory = scopeFactory;
+                }
+
+                public void Resolve(ref IMyService escaped)
+                {
+                    using var scope = _scopeFactory.CreateScope();
+            #pragma warning disable DI002
+                    var service = scope.ServiceProvider.GetRequiredService<IMyService>();
+            #pragma warning restore DI002
+                    escaped = service;
+                }
+            }
+
+            public class Startup
+            {
+                public void ConfigureServices(IServiceCollection services)
+                {
+                    services.AddScoped<IMyService, ScopedMyService>();
+                }
+            }
+
+            public class ScopedMyService : IMyService { }
+            """;
+
+        var expected = CodeFixVerifier<DI002_ScopeEscapeAnalyzer, DI002_ScopeEscapeCodeFixProvider>
+            .Diagnostic(DiagnosticDescriptors.ScopedServiceEscapes)
+            .WithLocation(0)
+            .WithArguments("escaped");
+
+        await CodeFixVerifier<DI002_ScopeEscapeAnalyzer, DI002_ScopeEscapeCodeFixProvider>
+            .VerifyCodeFixAsync(source, expected, fixedSource, "DI002_Suppress");
+    }
+
+    [Fact]
+    public async Task CodeFix_AddsPragmaSuppress_CapturedDelegateReturnEscape()
+    {
+        var source = Usings + """
+            public interface IMyService
+            {
+                void DoWork();
+            }
+
+            public class MyClass
+            {
+                private readonly IServiceScopeFactory _scopeFactory;
+
+                public MyClass(IServiceScopeFactory scopeFactory)
+                {
+                    _scopeFactory = scopeFactory;
+                }
+
+                public Action GetWork()
+                {
+                    Action work;
+                    using (var scope = _scopeFactory.CreateScope())
+                    {
+                        var service = {|#0:scope.ServiceProvider.GetRequiredService<IMyService>()|};
+                        work = () => service.DoWork();
+                    }
+
+                    return work;
+                }
+            }
+
+            public class Startup
+            {
+                public void ConfigureServices(IServiceCollection services)
+                {
+                    services.AddScoped<IMyService, ScopedMyService>();
+                }
+            }
+
+            public class ScopedMyService : IMyService
+            {
+                public void DoWork() { }
+            }
+            """;
+
+        var fixedSource = Usings + """
+            public interface IMyService
+            {
+                void DoWork();
+            }
+
+            public class MyClass
+            {
+                private readonly IServiceScopeFactory _scopeFactory;
+
+                public MyClass(IServiceScopeFactory scopeFactory)
+                {
+                    _scopeFactory = scopeFactory;
+                }
+
+                public Action GetWork()
+                {
+                    Action work;
+                    using (var scope = _scopeFactory.CreateScope())
+                    {
+            #pragma warning disable DI002
+                        var service = scope.ServiceProvider.GetRequiredService<IMyService>();
+            #pragma warning restore DI002
+                        work = () => service.DoWork();
+                    }
+
+                    return work;
+                }
+            }
+
+            public class Startup
+            {
+                public void ConfigureServices(IServiceCollection services)
+                {
+                    services.AddScoped<IMyService, ScopedMyService>();
+                }
+            }
+
+            public class ScopedMyService : IMyService
+            {
+                public void DoWork() { }
+            }
+            """;
+
+        var expected = CodeFixVerifier<DI002_ScopeEscapeAnalyzer, DI002_ScopeEscapeCodeFixProvider>
+            .Diagnostic(DiagnosticDescriptors.ScopedServiceEscapes)
+            .WithLocation(0)
+            .WithArguments("return");
 
         await CodeFixVerifier<DI002_ScopeEscapeAnalyzer, DI002_ScopeEscapeCodeFixProvider>
             .VerifyCodeFixAsync(source, expected, fixedSource, "DI002_Suppress");